chore: Move db endpoints into separste modules
