+++ .gitignore
@@ -11,8 +11,6 @@ logs
 node_modules
 
 # Generated files
-**/db/**
-**/db-test/**
 dist
 dist-ssr
 ./config
