+++ server/routes/v0/db/users/[uid]/Params.d.ts
@@ -1,0 +1,3 @@
+export interface Params {
+	uid: string;
+}
