+++ server/routes/v0/db/users/[uid]/[collectionId]/Params.d.ts
@@ -1,0 +1,4 @@
+export interface Params {
+	uid: string;
+	collectionId: string;
+}
