+++ server/routes/v0/db/users/[uid]/[collectionId]/[documentId]/Params.d.ts
@@ -1,0 +1,5 @@
+export interface Params {
+	uid: string;
+	collectionId: string;
+	documentId: string;
+}
