+++ server/routes/v0/db/users/[uid]/attachments/[documentId]/Params.d.ts
@@ -1,0 +1,4 @@
+export interface Params {
+	uid: string;
+	documentId: string;
+}
