+++ server/routes/v0/db/users/[uid]/attachments/[documentId]/blob/[fileName]/Params.d.ts
@@ -1,0 +1,5 @@
+export interface Params {
+	uid: string;
+	documentId: string;
+	fileName: string;
+}
